src/test/java/com/leetcode/linkedlists/medium/PartitionList_86Test.java
package com.leetcode.linkedlists.medium;

import com.leetcode.utils.LinkedList;
import com.leetcode.utils.LinkedList.ListNode;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static com.leetcode.assertions.LinkedListAssertions.assertThatLinkedList;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class PartitionList_86Test {

  private final PartitionList_86 solution = new PartitionList_86();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(LinkedList.of(1, 4, 3, 2, 5, 2), 3, LinkedList.of(1, 2, 2, 4, 3, 5)));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void partition(ListNode input, int x, ListNode expectedResult) {
    ListNode actualResult = solution.partition(input, x);

    assertThatLinkedList(actualResult).isEqualTo(expectedResult);
  }
}
